refactor cpu calculate

+++ types/container.go
@@ -16,7 +16,7 @@ type Container struct {
 	EntryPoint  string
 	Ident       string
 	Version     string
-	CPUNum      int
+	CPUNum      float64
 	CPUQuota    int64
 	CPUPeriod   int64
 	CPUShares   int64
@@ -27,17 +27,20 @@ type Container struct {
 	HealthCheck *coretypes.HealthCheck
 }
 
+// PrevCheck store healthcheck data
 type PrevCheck struct {
 	sync.Mutex
 	data map[string]bool
 }
 
+// Set set data to store
 func (p *PrevCheck) Set(ID string, f bool) {
 	p.Lock()
 	defer p.Unlock()
 	p.data[ID] = f
 }
 
+// Get get data from store
 func (p *PrevCheck) Get(ID string) bool {
 	p.Lock()
 	defer p.Unlock()
@@ -48,12 +51,14 @@ func (p *PrevCheck) Get(ID string) bool {
 	return v
 }
 
+// Del delete data from store
 func (p *PrevCheck) Del(ID string) {
 	p.Lock()
 	defer p.Unlock()
 	delete(p.data, ID)
 }
 
+// NewPrevCheck new a prevcheck obj
 func NewPrevCheck() *PrevCheck {
 	return &PrevCheck{
 		sync.Mutex{},
